Add helper functions and update data types

+++ src/cuda_compute.cu
@@ -84,7 +84,7 @@ __global__ void expKernel2d(T* data, T* out, size_t x, size_t y) {
 
 
 template <typename T>
-__global__ void powKernel(T* a, T *out, double n, int size){
+__global__ void powKernel(T* a, T *out, float n, int size){
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
         out[i] = pow(a[i], n);
@@ -92,7 +92,7 @@ __global__ void powKernel(T* a, T *out, double n, int size){
 }
 
 template <typename T> 
-__global__ void powKernel2d(T* a, T* out, double n, size_t x, size_t y) {
+__global__ void powKernel2d(T* a, T* out, float n, size_t x, size_t y) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     int j = blockIdx.y * blockDim.y + threadIdx.y;
     if (i < x && j < y) {
@@ -100,7 +100,7 @@ __global__ void powKernel2d(T* a, T* out, double n, size_t x, size_t y) {
     }
 }
 
-__global__ void cudaPowGrad(double* data, double *outGrad, double* grad, double n, int size){
+__global__ void cudaPowGrad(float* data, float *outGrad, float* grad, float n, int size){
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
         grad[i] += n * pow(data[i], n - 1) * outGrad[i];
@@ -111,7 +111,7 @@ template <typename T>
 __global__ void tanhKernel(T* data, T* out, int size){ 
     int i = blockIdx.x * blockDim.x + threadIdx.x; 
     if (i < size){
-        double x = data[i]; 
+        float x = data[i]; 
         out[i] = (exp(2.0 * x) - 1.0) / (exp(2.0 * x) + 1.0);
     }
 }
@@ -121,15 +121,15 @@ __global__ void tanhKernel2d(T* data, T* out, size_t x, size_t y) {
     int i = blockIdx.x * blockDim.x + threadIdx.x; 
     int j = blockIdx.y * blockDim.y + threadIdx.y;
     if (i < x && j < y){
-        double x = data[i * y + j]; 
+        float x = data[i * y + j]; 
         out[i * y + j] = (exp(2.0 * x) - 1.0) / (exp(2.0 * x) + 1.0);
     }    
 }
 
-__global__ void gradTanh(double* grad, double* outGrad, double* data, int size){
+__global__ void gradTanh(float* grad, float* outGrad, float* data, int size){
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
-        double t = (exp(2.0 * data[i]) - 1.0) / (exp(2.0 * data[i]) + 1.0);
+        float t = (exp(2.0 * data[i]) - 1.0) / (exp(2.0 * data[i]) + 1.0);
         grad[i] += (1.0 - pow(t, 2)) * outGrad[i];
     }
 }
@@ -151,7 +151,7 @@ __global__ void greaterKernel(T* a, T* b, T* c, int size) {
 }
 
 template <typename T>
-__global__ void greaterKernel(T* a, double b, T* c, int size) {
+__global__ void greaterKernel(T* a, float b, T* c, int size) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
         c[i] = a[i] > b;
@@ -168,7 +168,7 @@ __global__ void greaterKernel2d(T* a, T* b, T* c, size_t x, size_t y) {
 }
 
 template <typename T>
-__global__ void greaterKernel2d(T* a, double b, T* c, size_t x, size_t y) {
+__global__ void greaterKernel2d(T* a, float b, T* c, size_t x, size_t y) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     int j = blockIdx.y * blockDim.y + threadIdx.y;
     if (i < x && j < y) {
@@ -185,7 +185,7 @@ __global__ void lessKernel(T* a, T* b, T* c, int size) {
 }
 
 template <typename T> 
-__global__ void lessKernel(T* a, double b, T* c, int size) {
+__global__ void lessKernel(T* a, float b, T* c, int size) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
         c[i] = a[i] < b;
@@ -202,7 +202,7 @@ __global__ void lessKernel2d(T* a, T* b, T* c, size_t x, size_t y) {
 }
 
 template <typename T>
-__global__ void lessKernel2d(T* a, double b, T* c, size_t x, size_t y) {
+__global__ void lessKernel2d(T* a, float b, T* c, size_t x, size_t y) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     int j = blockIdx.y * blockDim.y + threadIdx.y;
     if (i < x && j < y) {
@@ -220,7 +220,7 @@ __global__ void equalKernel(T* a, T* b, T* c, int size) {
 
 
 template <typename T>
-__global__ void equalKernel(T* a, double b, T* c, int size) {
+__global__ void equalKernel(T* a, float b, T* c, int size) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
         c[i] = a[i] == b;
@@ -228,7 +228,7 @@ __global__ void equalKernel(T* a, double b, T* c, int size) {
 }
 
 template <typename T>
-__global__ void equalKernel2d(T* a, double b, T* c, size_t x, size_t y) {
+__global__ void equalKernel2d(T* a, float b, T* c, size_t x, size_t y) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     int j = blockIdx.y * blockDim.y + threadIdx.y;
     if (i < x && j < y) {
@@ -254,7 +254,7 @@ __global__ void greaterEqualKernel(T* a, T* b, T* c, int size) {
 }
 
 template <typename T>
-__global__ void greaterEqualKernel(T* a, double b, T* c, int size) {
+__global__ void greaterEqualKernel(T* a, float b, T* c, int size) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
         c[i] = a[i] >= b;
@@ -271,7 +271,7 @@ __global__ void greaterEqualKernel2d(T* a, T* b, T* c, size_t x, size_t y) {
 }
 
 template <typename T>
-__global__ void greaterEqualKernel2d(T* a, double b, T* c, size_t x, size_t y) {
+__global__ void greaterEqualKernel2d(T* a, float b, T* c, size_t x, size_t y) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     int j = blockIdx.y * blockDim.y + threadIdx.y;
     if (i < x && j < y) {
@@ -288,7 +288,7 @@ __global__ void lessEqualKernel(T* a, T* b, T* c, int size) {
 }
 
 template <typename T>
-__global__ void lessEqualKernel(T* a, double b, T* c, int size) {
+__global__ void lessEqualKernel(T* a, float b, T* c, int size) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
         c[i] = a[i] <= b;
@@ -305,7 +305,7 @@ __global__ void lessEqualKernel2d(T* a, T* b, T* c, size_t x, size_t y) {
 }
 
 template <typename T>
-__global__ void lessEqualKernel2d(T* a, double b, T* c, size_t x, size_t y) {
+__global__ void lessEqualKernel2d(T* a, float b, T* c, size_t x, size_t y) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     int j = blockIdx.y * blockDim.y + threadIdx.y;
     if (i < x && j < y) {
@@ -315,7 +315,7 @@ __global__ void lessEqualKernel2d(T* a, double b, T* c, size_t x, size_t y) {
 
 
 template <typename T>
-__global__ void addKernel(T* a, double b, T* c, int size) {
+__global__ void addKernel(T* a, float b, T* c, int size) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
         c[i] = a[i] + b;
@@ -332,7 +332,7 @@ __global__ void addKernel2d(T* a, T* b, T* c, size_t x, size_t y) {
 }
 
 template <typename T>
-__global__ void addKernel2d(T* a, double b, T* c, size_t x, size_t y) {
+__global__ void addKernel2d(T* a, float b, T* c, size_t x, size_t y) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     int j = blockIdx.y * blockDim.y + threadIdx.y;
     if (i < x && j < y) {
@@ -392,7 +392,7 @@ __global__ void sumKernel2daxis1(T* a, T* c, size_t x, size_t y){
     } 
 }
 
-__global__ void addGrad(double* grad, double* outGrad, double* otherGrad, int size) {
+__global__ void addGrad(float* grad, float* outGrad, float* otherGrad, int size) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
         grad[i] += 1.0 * outGrad[i];
@@ -410,7 +410,7 @@ __global__ void mulKernel(T* data, T* other, T* result, int size) {
 
 
 template <typename T> 
-__global__ void mulKernel(T* data, double other, T* result, int size) {
+__global__ void mulKernel(T* data, float other, T* result, int size) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
         result[i] = data[i] * other;
@@ -444,7 +444,7 @@ __global__ void mulKernel2dRowWise(T* data, T* other, T* result, size_t x, size_
     }
 }
 
-__global__ void mulKernel2dColWise(double* data, double* other, double* result, size_t x, size_t y) {
+__global__ void mulKernel2dColWise(float* data, float* other, float* result, size_t x, size_t y) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     int j = blockIdx.y * blockDim.y + threadIdx.y;
     if (i < x && j < y) {
@@ -453,7 +453,7 @@ __global__ void mulKernel2dColWise(double* data, double* other, double* result, 
 }
 
 template <typename T>
-__global__ void mulKernel2d(T* data, double other, T* result, size_t x, size_t y) {
+__global__ void mulKernel2d(T* data, float other, T* result, size_t x, size_t y) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     int j = blockIdx.y * blockDim.y + threadIdx.y;
     if (i < x && j < y) {
@@ -486,7 +486,7 @@ __global__ void dotKernel2d(T *a, T *b, T *result, size_t widthA, size_t heightA
 
 
 template<typename T>
-__global__ void dotKernel(T* data, double other, T* result, int size) {
+__global__ void dotKernel(T* data, float other, T* result, int size) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
         result[i] = data[i] * other;
@@ -534,10 +534,10 @@ __global__ void sigmoidKernel2d(T* data, T* out, size_t x, size_t y) {
 
 
 template <typename T>
-__global__ void gradSigmoid(double* grad, double* outGrad, double* data, int size) {
+__global__ void gradSigmoid(float* grad, float* outGrad, float* data, int size) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
-        double s = 1.0 / (1.0 + expf(-data[i]));
+        float s = 1.0 / (1.0 + expf(-data[i]));
         grad[i] += s * (1 - s) * outGrad[i];
     }
 }
@@ -560,7 +560,7 @@ __global__ void reluKernel2d(T* data, T* out, size_t x, size_t y) {
 }
 
 template <typename T>
-__global__ void gradRelu(double* grad, double* outGrad, double* data, int size) {
+__global__ void gradRelu(float* grad, float* outGrad, float* data, int size) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < size) {
         grad[i] += data[i] > 0 ? outGrad[i] : 0;
